requirements.txt
flask
edge-tts
oss2
python-dotenv